Refactor TeacherDB for Firebase
