+++ envolve-backend/src/main/java/de/neuefische/erbay/envolve/db/SchoolClassDb.java
@@ -10,7 +10,6 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 import java.util.Optional;
 import java.util.UUID;
-import java.util.concurrent.ExecutionException;
 import java.util.stream.Collectors;
 
 @Repository
